+++ lib/ui/BrowsePanel.dart
@@ -27,6 +27,7 @@ import 'package:illinois/ui/career/CareerPlanningLinks.dart';
 import 'package:illinois/ui/dining/DiningLinksPanel.dart';
 import 'package:illinois/ui/directory/DirectoryAccounts2Panel.dart';
 import 'package:illinois/ui/groups/GroupHome2Panel.dart';
+import 'package:illinois/ui/home/HomeSavedResourcesWidget.dart';
 import 'package:illinois/ui/illini/WordlePanel.dart';
 import 'package:illinois/ui/messages/MessagesHomePanel.dart';
 import 'package:illinois/ui/events2/Event2HomePanel.dart';
@@ -563,7 +564,15 @@ class _BrowseSection extends StatelessWidget {
       for(String code in _browseEntriesCodes!.reversed) {
         HomeFavorite? entryFavorite = _favorite(code);
         if (entryFavorite != null) {
-          favorites.add(entryFavorite);
+          String? entryCategory = entryFavorite.category;
+          Set<String>? entryCategoryItems = ((entryCategory != null) && (entryCategory != sectionId) && (_homeRootEntriesCodes?.contains(entryCategory) == true)) ?
+            _homeSectionsEntriesCodesMap[entryCategory] : null;
+          if (entryCategoryItems != null) {
+            favorites.add(HomeFavorite(entryCategory));
+            favorites.addAll(entryCategoryItems.map((itemId) => HomeFavorite(itemId, category: entryCategory)));
+          } else {
+            favorites.add(entryFavorite);
+          }
         }
       }
     }
@@ -888,8 +897,8 @@ class _BrowseEntry extends StatelessWidget {
   }
 
   static void _onTapMyCampusGuide(BuildContext context) {
-    Analytics().logSelect(target: "My Campus Guide");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) { return SavedPanel(favoriteCategories: [GuideFavorite.favoriteKeyName]); } ));
+    Analytics().logSelect(target: "Saved Resources");
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => SavedResourcesPanel()));
   }
 
   static void _onTapWellnessResources(BuildContext context) {