+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [8.2.9] - 2026-08-24
+### Fixed
+- Fixed crash in SwipeDetector widget [#6006](https://github.com/rokwire/illinois-app/issues/6006).
+### Changed
+- Reworked Saved Resources home widget and related functionality, implemented Saved Resources panel [#5873](https://github.com/rokwire/illinois-app/issues/5873).
+
 ## [8.2.8] - 2026-08-21
 ### Changed
 - Decreased chevron icon size in Home / Section browse entries [#5944](https://github.com/rokwire/illinois-app/issues/5944).